+++ source/scan-tool/Strategies.hpp
@@ -30,7 +30,7 @@ namespace virustotal
 {
 
 ///enumeration class for scan strategies
-enum class Strategy { None, Default };
+enum class Strategy { None, Default, DirectScan };
 
 
 /** \brief converts strategy to a simple string